Delete RBuds.Rproj
I think that deleting this .Rproj, together with adding *.Rproj to the .gitignore, should allow each of us to have our own preferred project settings. I likee 4 spaces for tabs, for example, but others may like something else. This may become an issue if we collaborate on the same documents, but we probably won't do too much of that.